add postgresql

+++ src/main/java/org/example/request/method/DeleteRequest.java
@@ -1,4 +1,4 @@
-package org.example.http_request.method;
+package org.example.request.method;
 
 import org.example.domain.User;
 import org.example.service.impl.UserServiceImpl;
@@ -17,14 +17,24 @@ public class DeleteRequest implements Request{
 
         User deletedUser = SERVICE.findById(id);
 
-        SERVICE.deleteById(id);
-
-        makeResponse(resp, logger, deletedUser);
+        if (deletedUser != null){
+            SERVICE.deleteById(id);
+            makeResponse(resp, logger, deletedUser);
+        } else {
+            makeResponse(resp, logger);
+        }
     }
 
     private void makeResponse(HttpServletResponse resp, Logger logger, User deletedUser) throws IOException {
         String msg = String.format("The user %s was deleted", deletedUser.getLogin());
         logger.error(msg);
         resp.getWriter().write(msg);
+    }
+
+    private void makeResponse(HttpServletResponse resp, Logger logger) throws IOException {
+        String msg = "There is no User with such id";
+        logger.error(msg);
+        resp.getWriter().write(msg);
+        resp.setStatus(404);
     }
 }
